pyproject.toml
[tool.poetry]
name = "diploma-goals-planner"
version = "0.1.0"
description = ""
authors = ["yasmirnov <iaroslav.irnovsm@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
envparse = "^0.2.0"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
social-auth-app-django = "^5.2.0"
django-filter = "^23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
